+++ R/reexport.R
@@ -23,9 +23,9 @@ dplyr::vars
 dplyr::select
 
 # huxtable ---------------------------------------------------------------------
-#' @export
-#' @importFrom huxtable as_huxtable
-huxtable::as_huxtable
+# #' @export
+# #' @importFrom huxtable as_huxtable
+# huxtable::as_huxtable
 
 # tidyselect -------------------------------------------------------------------
 #' @importFrom tidyselect starts_with