+++ tests/testthat/test-add_n.R
@@ -1,4 +1,5 @@
 context("test-add_n")
+testthat::skip_on_cran()
 
 t1 <- trial %>% tbl_summary()
 t2 <- trial %>% tbl_summary(by = trt)