+++ tests/testthat/test-add_overall.R
@@ -1,4 +1,5 @@
 context("test-add_overall")
+testthat::skip_on_cran()
 library(survival)
 
 test_that("no errors/warnings with standard use", {
@@ -22,5 +23,4 @@ test_that("no errors/warnings with missing data", {
 
 test_that("no errors/warnings with missing data in by variable", {
   expect_error(trial %>% tbl_summary(by = response) %>% add_overall(), NA)
-
 })