+++ tests/testthat/test-inline_text.R
@@ -1,4 +1,5 @@
 context("test-inline_text")
+testthat::skip_on_cran()
 
 
 # inline_text.tbl_summary tests --------------
@@ -82,11 +83,14 @@ test_that("inline_text.tbl_summary: with by -  expect errors", {
 })
 
 test_that("inline_text.tbl_summary: no errors with empty string selection", {
-  trial %>%
-    select(grade) %>%
-    mutate(grade = ifelse(grade == "I", "", as.character(grade))) %>%
-    tbl_summary() %>%
-    inline_text(variable = grade, level = "III")
+  expect_error(
+    trial %>%
+      select(grade) %>%
+      mutate(grade = ifelse(grade == "I", "", as.character(grade))) %>%
+      tbl_summary() %>%
+      inline_text(variable = grade, level = "III"),
+    NA
+  )
 })