+++ tests/testthat/test-tbl_summary_input_checks.R
@@ -1,4 +1,5 @@
 context("test-tbl_summary_input_checks")
+testthat::skip_on_cran()
 
 test_that("input check", {
   expect_error(