+++ NEWS.md
@@ -4,7 +4,9 @@
 
 * `tbl_stack()` now accepts a named list of gtsummary tables when using the `group_header=` argument (#524)
 
-* `add_p.tbl_summary()` bug fix for Wilcoxon rank-sum p-value calculation introduced in the last release (#525 )
+* `add_p.tbl_summary()` bug fix for Wilcoxon rank-sum p-value calculation introduced in the last release (#525)
+
+* Delaying the release of `as_huxtable()` until the next {huxtable} release.
 
 # gtsummary 1.3.1